+++ src/alloy_evm/mod.rs
@@ -1,1 +1,223 @@
-pub mod factory;
+use crate::transaction::POL_TX_TYPE;
+use alloy_primitives::{Address, Bytes, TxKind};
+use reth::revm::{
+    Context, ExecuteEvm, InspectEvm, InspectSystemCallEvm, Inspector, MainBuilder, MainContext,
+    SystemCallEvm,
+    context::{
+        BlockEnv, CfgEnv, Evm as RevmEvm, TxEnv,
+        result::{EVMError, HaltReason, ResultAndState},
+    },
+    handler::{EthFrame, EthPrecompiles, PrecompileProvider, instructions::EthInstructions},
+    inspector::NoOpInspector,
+    interpreter::{InterpreterResult, interpreter::EthInterpreter},
+    precompile::{PrecompileSpecId, Precompiles},
+    primitives::hardfork::SpecId,
+};
+use reth_evm::{
+    Database, Evm, EvmEnv, EvmFactory, eth::EthEvmContext, precompiles::PrecompilesMap,
+};
+use std::ops::{Deref, DerefMut};
+
+/// Berachain EVM implementation.
+///
+/// This is a wrapper type around the `revm` ethereum evm with optional [`Inspector`] (tracing)
+/// support. [`Inspector`] support is configurable at runtime because it's part of the underlying
+/// [`RevmEvm`] type.
+#[expect(missing_debug_implementations)]
+pub struct BerachainEvm<DB: Database, I, PRECOMPILE = EthPrecompiles> {
+    inner: RevmEvm<
+        EthEvmContext<DB>,
+        I,
+        EthInstructions<EthInterpreter, EthEvmContext<DB>>,
+        PRECOMPILE,
+        EthFrame,
+    >,
+    inspect: bool,
+}
+
+impl<DB: Database, I, PRECOMPILE> BerachainEvm<DB, I, PRECOMPILE> {
+    /// Creates a new Berachain EVM instance.
+    ///
+    /// The `inspect` argument determines whether the configured [`Inspector`] of the given
+    /// [`RevmEvm`] should be invoked on [`Evm::transact`].
+    pub const fn new(
+        evm: RevmEvm<
+            EthEvmContext<DB>,
+            I,
+            EthInstructions<EthInterpreter, EthEvmContext<DB>>,
+            PRECOMPILE,
+            EthFrame,
+        >,
+        inspect: bool,
+    ) -> Self {
+        Self { inner: evm, inspect }
+    }
+
+    /// Consumes self and return the inner EVM instance.
+    pub fn into_inner(
+        self,
+    ) -> RevmEvm<
+        EthEvmContext<DB>,
+        I,
+        EthInstructions<EthInterpreter, EthEvmContext<DB>>,
+        PRECOMPILE,
+        EthFrame,
+    > {
+        self.inner
+    }
+
+    /// Provides a reference to the EVM context.
+    pub const fn ctx(&self) -> &EthEvmContext<DB> {
+        &self.inner.ctx
+    }
+
+    /// Provides a mutable reference to the EVM context.
+    pub fn ctx_mut(&mut self) -> &mut EthEvmContext<DB> {
+        &mut self.inner.ctx
+    }
+}
+
+impl<DB: Database, I, PRECOMPILE> Deref for BerachainEvm<DB, I, PRECOMPILE> {
+    type Target = EthEvmContext<DB>;
+
+    #[inline]
+    fn deref(&self) -> &Self::Target {
+        self.ctx()
+    }
+}
+
+impl<DB: Database, I, PRECOMPILE> DerefMut for BerachainEvm<DB, I, PRECOMPILE> {
+    #[inline]
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        self.ctx_mut()
+    }
+}
+
+impl<DB, I, PRECOMPILE> BerachainEvm<DB, I, PRECOMPILE>
+where
+    DB: Database,
+    I: Inspector<EthEvmContext<DB>>,
+    PRECOMPILE: PrecompileProvider<EthEvmContext<DB>, Output = InterpreterResult>,
+{
+}
+
+impl<DB, I, PRECOMPILE> Evm for BerachainEvm<DB, I, PRECOMPILE>
+where
+    DB: Database,
+    I: Inspector<EthEvmContext<DB>>,
+    PRECOMPILE: PrecompileProvider<EthEvmContext<DB>, Output = InterpreterResult>,
+{
+    type DB = DB;
+    type Tx = TxEnv;
+    type Error = EVMError<DB::Error>;
+    type HaltReason = HaltReason;
+    type Spec = SpecId;
+    type Precompiles = PRECOMPILE;
+    type Inspector = I;
+
+    fn block(&self) -> &BlockEnv {
+        &self.block
+    }
+
+    fn chain_id(&self) -> u64 {
+        self.cfg.chain_id
+    }
+
+    fn transact_raw(
+        &mut self,
+        tx: Self::Tx,
+    ) -> Result<ResultAndState<Self::HaltReason>, Self::Error> {
+        if tx.tx_type == POL_TX_TYPE {
+            return match tx.kind {
+                TxKind::Create => Err(EVMError::Custom("pol tx cannot be create".into())),
+                TxKind::Call(to) => self.transact_system_call(tx.caller, to, tx.data),
+            };
+        }
+        if self.inspect { self.inner.inspect_tx(tx) } else { self.inner.transact(tx) }
+    }
+
+    fn transact_system_call(
+        &mut self,
+        caller: Address,
+        contract: Address,
+        data: Bytes,
+    ) -> Result<ResultAndState<Self::HaltReason>, Self::Error> {
+        if self.inspect {
+            self.inner.inspect_system_call_with_caller(caller, contract, data)
+        } else {
+            self.inner.system_call_with_caller(caller, contract, data)
+        }
+    }
+
+    fn finish(self) -> (Self::DB, EvmEnv<Self::Spec>) {
+        let Context { block: block_env, cfg: cfg_env, journaled_state, .. } = self.inner.ctx;
+
+        (journaled_state.database, EvmEnv { block_env, cfg_env })
+    }
+
+    fn set_inspector_enabled(&mut self, enabled: bool) {
+        self.inspect = enabled;
+    }
+
+    fn components(&self) -> (&Self::DB, &Self::Inspector, &Self::Precompiles) {
+        (&self.inner.ctx.journaled_state.database, &self.inner.inspector, &self.inner.precompiles)
+    }
+
+    fn components_mut(&mut self) -> (&mut Self::DB, &mut Self::Inspector, &mut Self::Precompiles) {
+        (
+            &mut self.inner.ctx.journaled_state.database,
+            &mut self.inner.inspector,
+            &mut self.inner.precompiles,
+        )
+    }
+}
+
+#[derive(Debug, Default, Clone, Copy)]
+#[non_exhaustive]
+pub struct BerachainEvmFactory;
+
+impl EvmFactory for BerachainEvmFactory {
+    type Evm<DB: Database, I: Inspector<EthEvmContext<DB>>> =
+        BerachainEvm<DB, I, Self::Precompiles>;
+    type Context<DB: Database> = Context<BlockEnv, TxEnv, CfgEnv, DB>;
+    type Tx = TxEnv;
+    type Error<DBError: core::error::Error + Send + Sync + 'static> = EVMError<DBError>;
+    type HaltReason = HaltReason;
+    type Spec = SpecId;
+    type Precompiles = PrecompilesMap;
+
+    fn create_evm<DB: Database>(&self, db: DB, input: EvmEnv) -> Self::Evm<DB, NoOpInspector> {
+        let spec_id = input.cfg_env.spec;
+        BerachainEvm {
+            inner: Context::mainnet()
+                .with_block(input.block_env)
+                .with_cfg(input.cfg_env)
+                .with_db(db)
+                .build_mainnet_with_inspector(NoOpInspector {})
+                .with_precompiles(PrecompilesMap::from_static(Precompiles::new(
+                    PrecompileSpecId::from_spec_id(spec_id),
+                ))),
+            inspect: false,
+        }
+    }
+
+    fn create_evm_with_inspector<DB: Database, I: Inspector<Self::Context<DB>>>(
+        &self,
+        db: DB,
+        input: EvmEnv,
+        inspector: I,
+    ) -> Self::Evm<DB, I> {
+        let spec_id = input.cfg_env.spec;
+        BerachainEvm {
+            inner: Context::mainnet()
+                .with_block(input.block_env)
+                .with_cfg(input.cfg_env)
+                .with_db(db)
+                .build_mainnet_with_inspector(inspector)
+                .with_precompiles(PrecompilesMap::from_static(Precompiles::new(
+                    PrecompileSpecId::from_spec_id(spec_id),
+                ))),
+            inspect: true,
+        }
+    }
+}
